requirements.txt
# Core libraries
numpy
pandas
matplotlib
seaborn
tqdm
scikit-image
jupyter

# Computer Vision
opencv-python
opencv-python-headless

# Machine Learning
tensorflow
keras
torch
torchvision
torchaudio
scikit-learn

# Web Application
streamlit
plotly
dash

# Image Processing
Pillow
pdf2image
PyMuPDF

# Database and Storage
sqlite3

# Data Export
openpyxl
xlsxwriter

# Additional utilities
python-dateutil
requests